Refactoring: gather cluster state management
code from Node into ClusterTopologyIntentTrackerImpl
