+++ hazelcast/src/main/java/com/hazelcast/instance/impl/Node.java
@@ -29,7 +29,6 @@ import com.hazelcast.config.Config;
 import com.hazelcast.config.DiscoveryConfig;
 import com.hazelcast.config.DiscoveryStrategyConfig;
 import com.hazelcast.config.EndpointConfig;
-import com.hazelcast.config.InvalidConfigurationException;
 import com.hazelcast.config.JoinConfig;
 import com.hazelcast.config.ListenerConfig;
 import com.hazelcast.config.MemberAttributeConfig;
@@ -101,7 +100,6 @@ import com.hazelcast.spi.impl.NodeEngineImpl;
 import com.hazelcast.spi.impl.proxyservice.impl.ProxyServiceImpl;
 import com.hazelcast.spi.properties.ClusterProperty;
 import com.hazelcast.spi.properties.HazelcastProperties;
-import com.hazelcast.spi.utils.RetryUtils;
 import com.hazelcast.version.MemberVersion;
 import com.hazelcast.version.Version;
 
@@ -115,11 +113,9 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static com.hazelcast.cluster.memberselector.MemberSelectors.DATA_MEMBER_SELECTOR;
 import static com.hazelcast.config.ConfigAccessor.getActiveMemberNetworkConfig;
@@ -139,7 +135,6 @@ import static com.hazelcast.spi.properties.ClusterProperty.DISCOVERY_SPI_PUBLIC_
 import static com.hazelcast.spi.properties.ClusterProperty.GRACEFUL_SHUTDOWN_MAX_WAIT;
 import static com.hazelcast.spi.properties.ClusterProperty.LOGGING_ENABLE_DETAILS;
 import static com.hazelcast.spi.properties.ClusterProperty.LOGGING_TYPE;
-import static com.hazelcast.spi.properties.ClusterProperty.PERSISTENCE_AUTO_CLUSTER_STATE_STRATEGY;
 import static com.hazelcast.spi.properties.ClusterProperty.SHUTDOWNHOOK_ENABLED;
 import static com.hazelcast.spi.properties.ClusterProperty.SHUTDOWNHOOK_POLICY;
 import static java.lang.Thread.currentThread;
@@ -147,7 +142,7 @@ import static java.security.AccessController.doPrivileged;
 
 @SuppressWarnings({"checkstyle:methodcount", "checkstyle:visibilitymodifier", "checkstyle:classdataabstractioncoupling",
         "checkstyle:classfanoutcomplexity"})
-public class Node implements ClusterTopologyIntentTracker {
+public class Node {
 
     public static final String DISCOVERY_PROPERTY_THIS_NODE = "hazelcast.internal.discovery.this.node";
 
@@ -187,12 +182,7 @@ public class Node implements ClusterTopologyIntentTracker {
     private final HealthMonitor healthMonitor;
     private final Joiner joiner;
     private final LocalAddressRegistry localAddressRegistry;
-    private final AtomicReference<ClusterTopologyIntent> clusterTopologyIntent =
-            new AtomicReference<>(ClusterTopologyIntent.NOT_IN_MANAGED_CONTEXT);
-    // single-threaded executor for actions in response to cluster topology intent changes
-    private final ExecutorService clusterTopologyExecutor;
-    // applies when automatic cluster state management is enabled (with persistence in kubernetes)
-    private final ClusterState clusterStateForMissingMembers;
+    private final ClusterTopologyIntentTracker clusterTopologyIntentTracker;
     private ManagementCenterService managementCenterService;
 
     // it can be changed on cluster service reset see: ClusterServiceImpl#resetLocalMemberUuid
@@ -220,11 +210,6 @@ public class Node implements ClusterTopologyIntentTracker {
         // Version can be overridden via system property. That's why BuildInfo should be parsed for each Node.
         this.buildInfo = BuildInfoProvider.getBuildInfo();
         this.version = MemberVersion.of(buildInfo.getVersion());
-        this.clusterStateForMissingMembers = properties.getEnum(PERSISTENCE_AUTO_CLUSTER_STATE_STRATEGY, ClusterState.class);
-        if (clusterStateForMissingMembers != ClusterState.FROZEN && clusterStateForMissingMembers != ClusterState.NO_MIGRATION) {
-            throw new InvalidConfigurationException("Value of property " + PERSISTENCE_AUTO_CLUSTER_STATE_STRATEGY.getName()
-                    + " was " + clusterStateForMissingMembers + " but should be one of FROZEN, NO_MIGRATION.");
-        }
 
         String loggingType = properties.getString(LOGGING_TYPE);
         boolean detailsEnabled = properties.getBoolean(LOGGING_ENABLE_DETAILS);
@@ -281,6 +266,12 @@ public class Node implements ClusterTopologyIntentTracker {
             healthMonitor = new HealthMonitor(this);
             clientEngine = hasClientServerSocket() ? new ClientEngineImpl(this) : new NoOpClientEngine();
             JoinConfig joinConfig = getActiveMemberNetworkConfig(this.config).getJoin();
+            if (properties.getBoolean(ClusterProperty.PERSISTENCE_AUTO_CLUSTER_STATE)
+                    && config.getPersistenceConfig().isEnabled()) {
+                clusterTopologyIntentTracker = new ClusterTopologyIntentTrackerImpl(this);
+            } else {
+                clusterTopologyIntentTracker = new NoOpClusterTopologyIntentTracker();
+            }
             DiscoveryConfig discoveryConfig = new DiscoveryConfigReadOnly(joinConfig.getDiscoveryConfig());
             List<DiscoveryStrategyConfig> aliasedDiscoveryConfigs =
                     AliasedDiscoveryConfigUtils.createDiscoveryStrategyConfigs(joinConfig);
@@ -294,7 +285,6 @@ public class Node implements ClusterTopologyIntentTracker {
             textCommandService = nodeExtension.createTextCommandService();
             multicastService = createMulticastService(addressPicker.getBindAddress(MEMBER), this, config, logger);
             joiner = nodeContext.createJoiner(this);
-            clusterTopologyExecutor = Executors.newSingleThreadExecutor();
         } catch (Throwable e) {
             try {
                 if (tmpLogger == null) {
@@ -353,9 +343,8 @@ public class Node implements ClusterTopologyIntentTracker {
         ILogger logger = getLogger(DiscoveryService.class);
 
         final Map attributes = new HashMap<>(localMember.getAttributes());
-        if (properties.getBoolean(ClusterProperty.PERSISTENCE_AUTO_CLUSTER_STATE)
-            && config.getPersistenceConfig().isEnabled()) {
-            attributes.put(DISCOVERY_PROPERTY_THIS_NODE, this);
+        if (clusterTopologyIntentTracker.isEnabled()) {
+            attributes.put(DISCOVERY_PROPERTY_THIS_NODE, clusterTopologyIntentTracker);
         }
         DiscoveryNode thisDiscoveryNode = new SimpleDiscoveryNode(localMember.getAddress(), attributes);
 
@@ -652,9 +641,6 @@ public class Node implements ClusterTopologyIntentTracker {
         if (healthMonitor != null) {
             healthMonitor.stop();
         }
-        if (clusterTopologyExecutor != null) {
-            clusterTopologyExecutor.shutdown();
-        }
     }
 
     private void mergeEnvironmentProvidedMemberMetadata() {
@@ -802,15 +788,16 @@ public class Node implements ClusterTopologyIntentTracker {
                 if (!isRunning()) {
                     return;
                 }
-                final ClusterTopologyIntent shutdownIntent = clusterTopologyIntent.get();
-                if (getNodeExtension().getInternalHotRestartService().isEnabled()
+                final ClusterTopologyIntent shutdownIntent = clusterTopologyIntentTracker.getClusterTopologyIntent();
+                if (clusterTopologyIntentTracker.isEnabled()
+                        && getNodeExtension().getInternalHotRestartService().isEnabled()
                         && shutdownIntent != ClusterTopologyIntent.UNKNOWN
                         && shutdownIntent != ClusterTopologyIntent.NOT_IN_MANAGED_CONTEXT) {
                     final ClusterState clusterState = clusterService.getClusterState();
                     logger.info("Running shutdown hook... Current node state: " + state
                                 + ", detected shutdown intent: " + shutdownIntent
                                 + ", cluster state: " + clusterState);
-                    shutdownWithIntent(shutdownIntent);
+                    clusterTopologyIntentTracker.shutdownWithIntent(shutdownIntent);
                 } else {
                     logger.info("Running shutdown hook... Current node state: " + state);
                 }
@@ -828,57 +815,31 @@ public class Node implements ClusterTopologyIntentTracker {
                 logger.warning(e);
             }
         }
-
-        private void shutdownWithIntent(ClusterTopologyIntent shutdownIntent) {
-            // consider the detected shutdown intent before triggering node shutdown
-            if (shutdownIntent == ClusterTopologyIntent.STABLE
-                || shutdownIntent == ClusterTopologyIntent.MISSING_MEMBERS) {
-                try {
-                    long timeoutNanos = getProperties().getNanos(ClusterProperty.CLUSTER_SHUTDOWN_TIMEOUT_SECONDS);
-                    while (!nodeEngine.getPartitionService().isPartitionTableHealthy()
-                            && timeoutNanos > 0) {
-                        sleep(TimeUnit.SECONDS.toMillis(1));
-                        timeoutNanos -= TimeUnit.SECONDS.toNanos(1);
-                    }
-                    changeClusterState(clusterStateForMissingMembers);
-                } catch (Throwable t) {
-                    // let shutdown proceed even though we failed to switch to FROZEN state
-                    logger.warning("Could not switch to transient FROZEN state while cluster"
-                            + "shutdown intent was " + shutdownIntent, t);
-                }
-            } else if (shutdownIntent == ClusterTopologyIntent.CLUSTER_SHUTDOWN
-                    || shutdownIntent == ClusterTopologyIntent.CLUSTER_SHUTDOWN_WITH_MISSING_MEMBERS) {
-                try {
-                    changeClusterState(ClusterState.PASSIVE);
-                } catch (Throwable t) {
-                    // let shutdown proceed even though we failed to switch to PASSIVE state
-                    // and wait for replica sync
-                    logger.warning("Could not switch to transient PASSIVE state while cluster"
-                            + "shutdown intent was " + shutdownIntent, t);
-                }
-                long timeoutNanos = getProperties().getNanos(ClusterProperty.CLUSTER_SHUTDOWN_TIMEOUT_SECONDS);
-                try {
-                    getNodeExtension().getInternalHotRestartService()
-                            .waitPartitionReplicaSyncOnCluster(timeoutNanos, TimeUnit.NANOSECONDS);
-                } catch (IllegalStateException e) {
-                    logger.severe("Failure while waiting for partition replica sync before shutdown", e);
-                }
-            }
-        }
     }
 
     public ClusterTopologyIntent getClusterTopologyIntent() {
-        return clusterTopologyIntent.get();
+        return clusterTopologyIntentTracker.getClusterTopologyIntent();
     }
 
     public void initializeClusterTopologyIntent(ClusterTopologyIntent clusterTopologyIntent) {
-        ClusterTopologyIntent current = this.clusterTopologyIntent.get();
-        logger.info("Current node cluster topology intent is " + current);
-        // if not UNKNOWN, then it was already initialized
-        if (current == ClusterTopologyIntent.UNKNOWN) {
-            logger.info("Initializing this node's cluster topology to " + clusterTopologyIntent);
-            this.clusterTopologyIntent.set(clusterTopologyIntent);
-        }
+        clusterTopologyIntentTracker.initializeClusterTopologyIntent(clusterTopologyIntent);
+    }
+
+    // true when Hazelcast is running in managed context with a tracker enabled
+    public boolean isClusterStateManagementAutomatic() {
+        return clusterTopologyIntentTracker.isEnabled();
+    }
+
+    /**
+     * When running in managed context with automatic cluster state management,
+     * return true if the current cluster size (as observed by Hazelcast's cluster service)
+     * is same as the cluster specification size (as provided by the runtime environment ie Kubernetes).
+     * @return  {@code true} if running with auto cluster state management and current cluster size is same as
+     *          specified one, otherwise {@code false}.
+     */
+    public boolean isClusterComplete() {
+        return clusterTopologyIntentTracker.isEnabled()
+                && clusterTopologyIntentTracker.getCurrentClusterSpecSize() == clusterService.getSize();
     }
 
     public SplitBrainJoinMessage createSplitBrainJoinMessage() {
@@ -1032,92 +993,5 @@ public class Node implements ClusterTopologyIntentTracker {
             }
         }
         return attributes;
-    }
-
-    @SuppressWarnings({"checkstyle:npathcomplexity", "checkstyle:cyclomaticcomplexity"})
-    @Override
-    public void update(int previousClusterSpecSize, int currentClusterSpecSize,
-                       int readyNodesCount, int currentNodesCount) {
-        if (previousClusterSpecSize == UNKNOWN) {
-            if (currentClusterSpecSize > 0
-                    && (readyNodesCount == UNKNOWN || readyNodesCount == 0)) {
-                // startup of first member of new cluster
-                logger.info("Cluster starting in managed context");
-                clusterTopologyIntent.set(ClusterTopologyIntent.CLUSTER_START);
-            } else {
-                logger.info("Member starting in managed context");
-                clusterTopologyIntent.set(ClusterTopologyIntent.UNKNOWN);
-            }
-            return;
-        }
-        final ClusterTopologyIntent previous = clusterTopologyIntent.get();
-        ClusterTopologyIntent newTopologyIntent;
-        if (currentClusterSpecSize == 0) {
-            newTopologyIntent = previous == ClusterTopologyIntent.MISSING_MEMBERS
-                        || previous == ClusterTopologyIntent.CLUSTER_SHUTDOWN_WITH_MISSING_MEMBERS
-                    ? ClusterTopologyIntent.CLUSTER_SHUTDOWN_WITH_MISSING_MEMBERS
-                    : ClusterTopologyIntent.CLUSTER_SHUTDOWN;
-        } else if (previousClusterSpecSize == currentClusterSpecSize) {
-            if (previous == ClusterTopologyIntent.SCALING
-                || previous == ClusterTopologyIntent.UNKNOWN) {
-                // only switch to STABLE when ready nodes count equals number of currentClusterSpecSize
-                if (readyNodesCount != currentClusterSpecSize) {
-                    logger.info("Ignoring state change because readyNodesCount "
-                        + readyNodesCount + ", while spec requires " + currentClusterSpecSize);
-                    return;
-                }
-            }
-            if (previous == ClusterTopologyIntent.CLUSTER_START
-                && readyNodesCount < currentClusterSpecSize) {
-                // cluster start is not done yet, don't switch to STABLE
-                logger.info("Ignoring state change because readyNodesCount "
-                        + readyNodesCount + " is less than required by spec and cluster is still starting");
-                return;
-            }
-            newTopologyIntent = (readyNodesCount < currentClusterSpecSize) || (currentNodesCount < currentClusterSpecSize)
-                ? ClusterTopologyIntent.MISSING_MEMBERS : ClusterTopologyIntent.STABLE;
-        } else {
-            newTopologyIntent = ClusterTopologyIntent.SCALING;
-        }
-        if (clusterTopologyIntent.compareAndSet(previous, newTopologyIntent)) {
-            logger.info("Cluster topology intent: " + previous + " -> " + newTopologyIntent);
-            clusterTopologyExecutor.submit(() -> {
-                nodeExtension.getInternalHotRestartService().onClusterTopologyIntentChange();
-                if (!isMaster()) {
-                    return;
-                }
-                if (newTopologyIntent == ClusterTopologyIntent.SCALING) {
-                    changeClusterState(ClusterState.ACTIVE);
-                } else if (newTopologyIntent == ClusterTopologyIntent.STABLE) {
-                    if (clusterService.getClusterState() != ClusterState.ACTIVE) {
-                        executeOrScheduleClusterStateChange(ClusterState.ACTIVE);
-                    } else if (!partitionService.isPartitionTableHealthy()) {
-                        partitionService.getMigrationManager().triggerControlTask();
-                    }
-                }
-            });
-        }
-    }
-
-    private void executeOrScheduleClusterStateChange(ClusterState newClusterState) {
-        if (!getNodeExtension().getInternalHotRestartService().setDeferredClusterState(newClusterState)) {
-            // hot restart recovery is completed, just apply the new cluster state here
-            changeClusterState(newClusterState);
-        }
-    }
-
-    /**
-     * Change cluster state, if current state is not already the desired one.
-     * Retries up to 3 times. The cluster state change is transient, so if persistence
-     * is enabled, the new cluster state is not persisted to disk.
-     *
-     * @param newClusterState
-     */
-    private void changeClusterState(ClusterState newClusterState) {
-        RetryUtils.retry(
-                () -> {
-                    clusterService.changeClusterState(newClusterState, true);
-                    return null;
-                }, 3);
     }
 }
